+++ src/candlewick/posteffects/SSAO.cpp
@@ -3,7 +3,7 @@
 #include "../core/CommandBuffer.h"
 #include "../core/Shader.h"
 #include "../core/Camera.h"
-#include "../core/Renderer.h"
+#include "../core/RenderContext.h"
 #include "../third-party/float16_t.hpp"
 #include <random>