+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- core : rename header `{ Renderer.h => RenderContext.h }`
+
 ## [0.2.0] - 2025-05-31
 
 ### Fixed